refactor(cockpit): extract _InboxProjectPickerModal into sibling module (#1354)
First wedge of the cockpit_ui.py god-module split. Moves the
~110-LOC inbox project-picker modal into cockpit_inbox_project_picker.py;
cockpit_ui re-exports the symbol so existing import paths and the
single call site keep working unchanged.

cockpit_ui.py: 18481 -> 18370 LOC.
